Fix all 118 compile errors - CombatManager, SectorMap, AbilityExecutor, PerformanceManager, EncounterGenerator, PlayerController, TutorialZone, ConsoleCertification
- Fixed Vector2I.Normalized() operator precedence in position effects
- Fixed CombatManager.ApplyPositionEffect signature and SpawnUnit visibility
- Fixed AbilityExecutor CombatAction namespace conflict (removed local class)
- Fixed SectorMap ZoneResource iteration (kvp.Key/kvp.Value)
- Fixed PerformanceManager Godot.Performance.Monitor enum values (MemoryStatic, RenderTotalDrawCallsInFrame)
- Fixed EncounterGenerator EnemySpawnInfo.Position → GridPosition
- Fixed PlayerController DerivedStats → inline computation
- Fixed TutorialZone GameManager.LoadScene → GetTree().ChangeSceneToFile
- Fixed ConsoleCertification missing NotificationType using
- Removed ZoneNode ZoneHovered/ZoneSelected event subscriptions (not in ZoneNode)
- Updated DialogueManager to use real YarnSpinnerGodot API

+++ Systems/NemesisSystem.cs
@@ -117,7 +117,7 @@ public partial class NemesisSystem : Node
 
         if (candidates.Count == 0) return null;
 
-        var nemesis = candidates[GD.Randi() % candidates.Count];
+        var nemesis = candidates[(int)(GD.Randi() % candidates.Count)];
 
         // Update tracking
         nemesis.LastEncounterZone = zoneId;
